+++ .meteor/versions
@@ -9,8 +9,6 @@ aldeed:collection2-core@1.1.1
 aldeed:schema-deny@1.0.1
 aldeed:schema-index@1.0.1
 aldeed:simple-schema@1.5.3
-aldeed:tabular@1.4.2
-aldeed:template-extension@4.0.0
 allow-deny@1.0.5
 audit-argument-checks@1.0.7
 autoupdate@1.2.10
@@ -50,7 +48,6 @@ cfs:ui@0.1.3
 cfs:upload-http@0.0.20
 cfs:worker@0.1.4
 check@1.2.2
-chuangbo:cookie@1.1.0
 coffeescript@1.1.0
 dburles:collection-helpers@1.0.4
 ddp@1.2.5
@@ -74,36 +71,20 @@ html-tools@1.0.10
 htmljs@1.0.10
 http@1.1.6
 id-map@1.0.8
-iron:controller@1.0.12
-iron:core@1.0.11
-iron:dynamic-template@1.0.12
-iron:layout@1.0.12
-iron:location@1.0.11
-iron:middleware-stack@1.1.0
-iron:router@1.0.13
-iron:url@1.0.11
 jquery@1.11.9
 launch-screen@1.0.12
 less@2.6.1
 livedata@1.0.18
 localstorage@1.0.11
 logging@1.0.13
-manuelschoebel:ms-seo@0.4.1
 matteodem:easy-security@0.1.5
 mdg:validation-error@0.2.0
 meteor@1.1.15
 meteor-base@1.0.4
-meteorhacks:fast-render@2.14.0
-meteorhacks:inject-data@2.0.0
-meteorhacks:meteorx@1.4.1
-meteorhacks:picker@1.0.3
-meteorhacks:unblock@1.1.0
-mfactory:admin-lte@0.0.2
 minifier-js@1.1.12
 minifiers@1.1.7
 minimongo@1.0.17
 mizzao:bootboxjs@4.4.0
-mizzao:bootstrap-3@3.3.1_1
 mobile-experience@1.0.4
 mobile-status-bar@1.0.12
 modules@0.6.2
@@ -130,7 +111,6 @@ reactive-dict@1.1.8
 reactive-var@1.0.10
 reload@1.1.9
 retry@1.0.8
-reywood:publish-composite@1.4.2
 routepolicy@1.0.11
 seba:minifiers-autoprefixer@0.0.2
 service-configuration@1.0.10
@@ -146,12 +126,8 @@ tracker@1.0.14
 twbs:bootstrap@3.3.6
 ui@1.0.11
 underscore@1.0.9
-underscorestring:underscore.string@3.0.3_1
 url@1.0.10
 vsivsi:file-collection@1.3.5
 webapp@1.2.9
 webapp-hashing@1.0.9
-yogiben:admin@1.2.6
 yogiben:autoform-file@0.4.2
-zimme:active-route@2.3.2
-zimme:iron-router-active@1.0.4
