--- client/compatibility/default.js
@@ -1,5 +1,0 @@
-/*
-// Libraries which need a global variable, such as jQuery etc
-globalVariable = function () {
-}
-*/
