--- client/config/router.js
@@ -1,3 +1,0 @@
-Router.configure({
-  layoutTemplate: 'basicLayout'
-});
