--- client/config/seo.js
@@ -1,6 +1,0 @@
-SEO.config({
-  title: Meteor.App.NAME,
-  meta: {
-    'description': Meteor.App.DESCRIPTION
-  }
-});
