+++ client/startup/default.js
@@ -4,6 +4,5 @@ import {render} from 'react-dom';
 import HomePage from '/imports/ui/pages/home';
 
 Meteor.startup(function () {
-  console.log('hello');
   render(<HomePage />, document.getElementById('react-root'));
 });
