+++ client/views/home/_homeHeader.html
@@ -34,4 +34,3 @@
     <div class="stars-underlay"></div>
   </header>
 </template>
-
